+++ order/src/main/java/com/smore/order/application/event/inbound/CompletedRefundEvent.java
@@ -1,7 +1,5 @@
-package com.smore.order.domain.event;
+package com.smore.order.application.event.inbound;
 
-import com.smore.order.domain.status.RefundStatus;
-import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
