+++ order/src/main/java/com/smore/order/application/event/outbound/OrderEvent.java
@@ -1,0 +1,5 @@
+package com.smore.order.application.event.outbound;
+
+public interface OrderEvent {
+
+}
