+++ order/src/main/java/com/smore/order/application/event/outbound/RefundFailedEvent.java
@@ -1,4 +1,4 @@
-package com.smore.order.domain.event;
+package com.smore.order.application.event.outbound;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
