+++ order/src/main/java/com/smore/order/application/listener/EventListener.java
@@ -5,10 +5,10 @@ import com.smore.order.application.dto.CompletedRefundCommand;
 import com.smore.order.application.dto.CreateOrderCommand;
 import com.smore.order.application.dto.FailedRefundCommand;
 import com.smore.order.application.service.OrderService;
-import com.smore.order.domain.event.BidRequestEvent;
-import com.smore.order.domain.event.CompletedPaymentEvent;
-import com.smore.order.domain.event.CompletedRefundEvent;
-import com.smore.order.domain.event.FailedRefundEvent;
+import com.smore.order.application.event.inbound.BidRequestEvent;
+import com.smore.order.application.event.inbound.CompletedPaymentEvent;
+import com.smore.order.application.event.inbound.CompletedRefundEvent;
+import com.smore.order.application.event.inbound.FailedRefundEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
