+++ order/src/main/java/com/smore/order/application/service/OrderService.java
@@ -11,12 +11,12 @@ import com.smore.order.application.exception.RefundReservationConflictException;
 import com.smore.order.application.repository.OrderRepository;
 import com.smore.order.application.repository.OutboxRepository;
 import com.smore.order.application.repository.RefundRepository;
-import com.smore.order.domain.event.CompletedOrderEvent;
-import com.smore.order.domain.event.CreatedOrderEvent;
-import com.smore.order.domain.event.OrderEvent;
-import com.smore.order.domain.event.RefundFailedEvent;
-import com.smore.order.domain.event.RefundRequestEvent;
-import com.smore.order.domain.event.SuccessRefundEvent;
+import com.smore.order.application.event.outbound.CompletedOrderEvent;
+import com.smore.order.application.event.outbound.CreatedOrderEvent;
+import com.smore.order.application.event.outbound.OrderEvent;
+import com.smore.order.application.event.outbound.RefundFailedEvent;
+import com.smore.order.application.event.outbound.RefundRequestEvent;
+import com.smore.order.application.event.outbound.SuccessRefundEvent;
 import com.smore.order.domain.model.Order;
 import com.smore.order.domain.model.Outbox;
 import com.smore.order.domain.model.Refund;
