--- order/src/main/java/com/smore/order/domain/event/OrderEvent.java
@@ -1,5 +1,0 @@
-package com.smore.order.domain.event;
-
-public interface OrderEvent {
-
-}
